+++ core/XH.ts
@@ -29,9 +29,11 @@ import {
     PrefService,
     SessionStorageService,
     TrackService,
+    TraceService,
     WebSocketService
 } from '@xh/hoist/svc';
 import {apiDeprecated, getLogLevel, LogLevel, setLogLevel} from '@xh/hoist/utils/js';
+import {SpanConfig} from '@xh/hoist/utils/telemetry';
 import {camelCase, flatten, isString, uniqueId} from 'lodash';
 import {Router, State} from 'router5';
 import {CancelFn} from 'router5/types/types/base';
@@ -157,6 +159,7 @@ export class XHApi {
     prefService: PrefService;
     sessionStorageService: SessionStorageService;
     trackService: TrackService;
+    traceService: TraceService;
     webSocketService: WebSocketService;
 
     //----------------------------
@@ -334,6 +337,15 @@ export class XHApi {
      */
     track(opts: string | TrackOptions) {
         return this.trackService?.track(opts);
+    }
+
+    /**
+     * Create a span wrapping an async operation.
+     * @see TraceService.withSpanAsync
+     */
+    async withSpanAsync<T>(config: string | SpanConfig, fn: (span) => Promise<T>): Promise<T> {
+        const svc = this.traceService;
+        return svc ? svc.withSpanAsync(config, fn) : fn(null);
     }
 
     /**